Move db and table creation to seperate file, add script to run
