+++ database-mysql/generateData.js
@@ -1,0 +1,8 @@
+const faker = require('faker');
+
+
+const generateLocations = function () {
+
+};
+
+exports.populateDb;
