+++ database-mysql/index.js
@@ -1,5 +1,8 @@
-var mysql = require('mysql');
-var connectionOptions = {
+const mysql = require('mysql');
+// const populate = require('./generateData.js');
+
+// ***********  db connection  ***********
+const connectionOptions = {
   user: 'root',
   password: ''
 }
@@ -14,114 +17,6 @@ dbConnection.connect((err) => {
   console.log('connected to mysql');
 });
 
-dbConnection.query(
-  'CREATE DATABASE IF NOT EXISTS products',
-  function (err) {
-    if (err) {
-      console.error('db creation error', err);
-    }
-  }
-);
 
-dbConnection.query('USE products');
-
-dbConnection.query(
-  `CREATE TABLE IF NOT EXISTS shipping (
-    id        INT       NOT NULL AUTO_INCREMENT,
-    type      CHAR(13)  NOT NULL,
-    free      BOOLEAN   NOT NULL,
-    timeframe CHAR(30),
-    PRIMARY KEY (id)
-  )`,
-  function (err) {
-    if (err) {
-      console.error('shipping table creation error', err);
-    }
-  }
-);
-
-dbConnection.query(
-  `CREATE TABLE IF NOT EXISTS locations (
-    id          INT       NOT NULL AUTO_INCREMENT,
-    country     CHAR(50)  NOT NULL,
-    state       CHAR(50),
-    city        CHAR(50),
-    PRIMARY KEY (id)
-  )`,
-  function (err) {
-    if (err) {
-      console.error('loations table creation error', err);
-    }
-  }
-);
-
-dbConnection.query(
-  `CREATE TABLE IF NOT EXISTS items (
-    id                INT           NOT NULL AUTO_INCREMENT,
-    title             CHAR(180)     NOT NULL,
-    price             FLOAT(7, 2)   NOT NULL,
-    shipping_id       INT           NOT NULL,
-    materials         CHAR(180)     NOT NULL,
-    description       TEXT          NOT NULL,
-    location_id       INT           NOT NULL,
-    policies          TEXT          NOT NULL,
-    return_synopsis   CHAR(100)     NOT NULL,
-    dimensions        VARCHAR(100),
-    max_order_qty     INT,
-    returns_condttion TEXT,
-    inventory_count   INT,
-    in_other_carts    INT,
-    gift_wrap         BOOLEAN,
-    faqs              BOOLEAN,
-    bestseller        BOOLEAN,
-    personalizable    BOOLEAN,
-    handmade          BOOLEAN,
-    vintage           BOOLEAN,
-    PRIMARY KEY (id),
-    FOREIGN KEY (shipping_id)
-      REFERENCES shipping(id),
-    FOREIGN KEY (location_id)
-      REFERENCES locations(id)
-  )`,
-  function (err) {
-    if (err) {
-      console.error('items table creation error', err);
-    }
-  }
-);
-
-dbConnection.query(
-  `CREATE TABLE IF NOT EXISTS options (
-    id        INT       NOT NULL AUTO_INCREMENT,
-    item_id   INT       NOT NULL,
-    title     CHAR(20)  NOT NULL,
-    list      TEXT      NOT NULL,
-    PRIMARY KEY (id),
-    FOREIGN KEY (item_id)
-      REFERENCES items(id)
-  )`,
-  function (err) {
-    if (err) {
-      console.error('options table creation error', err);
-    }
-  }
-);
-
-dbConnection.query(
-  `CREATE TABLE IF NOT EXISTS sales (
-    id          INT   NOT NULL AUTO_INCREMENT,
-    item_id     INT   NOT NULL,
-    discount    INT   NOT NULL,
-    end_date    date,
-    PRIMARY KEY (id),
-    FOREIGN KEY (item_id)
-      REFERENCES items(id)
-  )`,
-  function (err) {
-    if (err) {
-      console.error('sales table creation error', err);
-    }
-  }
-);
 
 exports.dbConnection;
